requirements.txt
# Backend Dependencies
fastapi==0.128.6
uvicorn[standard]==0.40.0
sqlalchemy==2.0.46
aiosqlite==0.22.1
pandas==3.0.0
numpy==2.4.2
httpx==0.28.1
python-dotenv==1.2.1
pydantic==2.12.5
pywebview==6.1
ib_insync==0.9.86
cryptography==46.0.5

# Build & Update Tools
pyinstaller==6.12.0
pyupdater==4.1.2
pyupdater-s3-plugin==0.1.0  # Optional, but commonly used
